third_party/WebKit/LayoutTests/css3/device-adapt/opera/constrain-003-expected.txt


FAIL CSS Test: @viewport constrained - min-width 10px assert_equals: expected 320 but got 980
